fix(engine): More robust tracking for Entity validation (#1301)
# Entity validation rework

This PR aims to build a robust system for how references to entities are
validated to be active entities in the game world. This applies to Locs,
Players, Objs, and Npcs

## New boolean Entity.isActive

`Entity` now has a new boolean `isActive`. `isActive` will be what we
use to keep track of if the `Entity` currently should be available in
the `World`

- `isActive` is set to `true` inside of `Zone.addObj()`,
`Zone.addStaticObj()`, `Zone.addLoc()`, and `Zone.addStaticLoc()` for
non-pathing entities
- `isActive` is also set to `true` inside of `World.addNpc()` and
`World.addPlayer()` for pathing entities
- `isActive` is set to `false` inside of `Zone.removeObj()`, and
`Zone.removeLoc()` non-pathing entities
- `isActive` is also set to `false` inside of `World.removeNpc()` and
`World.removePlayer()` for pathing entities

Essentially, whenever we add an `Entity` to the `World`, we mark the
`Entity` as `isActive`, and when we remove it from the `World`, we unset
`isActive`. This gives us a really quick way to check if the `Entity`
exists in the `World`

## New Boolean Obj.isRevealed

`Obj` also has a unique new boolean called `isRevealed`. `isRevealed` is
used to keep track of whether or not the `Obj` is available to all
players or just its initial owner

- `isRevealed` is set to `true` when `Zone.revealObj()` is called
- `isRevealed` is also set to `true` when `Zone.addStaticObj()` is
called

Inside of `Zone.addObj()`,

- `isRevealed` is set to `true` if the `Obj` has a lifecycle of
'RESPAWN' or is the `Obj` has no receiver
- `isRevealed` is set to `false` if the `Obj` has a receiver and a
lifecycle of 'DESPAWN'

## New polymorphic method Entity.isValid(): boolean

`Entity` has a new method `isValid(): boolean` that should be used to be
able to quickly check if an `Entity` is valid for interactions. This
method is to be overwritten by children classes for specific checks
related to each child

- `Entity.isValid()` by default just returns `isActive`
- `Loc.isValid()` does not overwrite `Entity`'s implementation
- `Npc.isValid()` first checks if the `Npc` is delayed, and then calls
`super.isEntity()`
- `Obj.isValid(hash64: bigint)` takes in the player's 64bit id and does
the following checks
- If the `Obj.isRevealed` is false, it checks that its receiver hash
matches `hash64`
  - It checks that `Obj.count` is greater than 0
  - Finally call `super.isEntity()`
- `Player.isValid()` does the following checks
  - Check `Player.loggingOut`
  - Make sure the player is visible
  - Call `super.isEntity()`

## Places where Entity.isValid() is being referenced for validation

- `PathingEntity.setInteraction()` now checks if the target is
`isValid()` before allowing interaction to be set
- `Player.validateTarget()` now checks if the target `isValid()` as one
of its checks
- `OBJ_COUNT` and `OBJ_TAKEITEM` commands now check against
`Obj.isValid(Player.hash64)`

## Issues this fixes

- https://github.com/2004Scape/Server/issues/1188
- https://discord.com/channels/953326730632904844/1340360421529616495
- Others